fake-news-detector/requirements.txt
pandas==2.2.2
numpy==2.0.1
scikit-learn==1.5.1
spacy==3.7.4
joblib==1.4.2
pyyaml==6.0.2
flask==3.0.3
streamlit==1.38.0
matplotlib==3.9.2
seaborn==0.13.2
lime==0.2.0.1
transformers==4.44.2
torch==2.4.0